пофиксил код
