+++ first-web-app/src/main/java/ru/geekbrains/filter/EncodingFilter.java
@@ -18,7 +18,7 @@ public class EncodingFilter implements Filter {
     public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain) throws IOException, ServletException {
         response.setContentType("text/html");
         response.setCharacterEncoding("UTF-8");
-
+        request.setCharacterEncoding("UTF-8");
         chain.doFilter(request, response);
     }
