--- first-web-app/src/main/java/ru/geekbrains/model/Product.java
@@ -1,56 +1,0 @@
-package ru.geekbrains.model;
-
-import java.math.BigDecimal;
-
-public class Product {
-
-    private Long id;
-
-    private String name;
-
-    private String description;
-
-    private BigDecimal price;
-
-    public Product() {
-    }
-
-    public Product(Long id, String name, String description, BigDecimal price) {
-        this.id = id;
-        this.name = name;
-        this.description = description;
-        this.price = price;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public BigDecimal getPrice() {
-        return price;
-    }
-
-    public void setPrice(BigDecimal price) {
-        this.price = price;
-    }
-}
