+++ first-web-app/src/main/java/ru/geekbrains/persist/Category.java
@@ -1,4 +1,4 @@
-package ru.geekbrains.model;
+package ru.geekbrains.persist;
 
 public class Category {
