+++ first-web-app/src/main/java/ru/geekbrains/persist/User.java
@@ -1,4 +1,4 @@
-package ru.geekbrains.model;
+package ru.geekbrains.persist;
 
 public class User {
