+++ first-web-app/src/main/java/ru/geekbrains/servlet/PageHeaderServlet.java
@@ -6,12 +6,15 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.io.PrintWriter;
 
-@WebServlet(urlPatterns = "/catalog")
-public class CatalogServlet extends HttpServlet {
+@WebServlet(name = "header", urlPatterns = "/header")
+public class PageHeaderServlet extends HttpServlet {
+
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        getServletContext().getRequestDispatcher("/header").include(req, resp);
-        resp.getWriter().println("<h1>Catalog page</h1>");
+        PrintWriter out = resp.getWriter();
+        out.close();
+        req.getRequestDispatcher("/header.html").include(req, resp);
     }
 }
