--- first-web-app/src/main/java/ru/geekbrains/servlet/PageProductServlet.java
@@ -1,18 +1,0 @@
-package ru.geekbrains.servlet;
-
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-@WebServlet(urlPatterns = "/products")
-public class PageProductServlet extends HttpServlet {
-
-    @Override
-    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        String header = (String) req.getAttribute("products");
-        resp.getWriter().println("<h1>" + header +"</h1>");
-    }
-}
