+++ first-web-app/src/main/webapp/header.html
@@ -1,0 +1,11 @@
+<div style="display:flex; background-color: aliceblue; font-size: 32px; height: 40px; width: 60%;
+padding-top:16px; margin: auto" >
+    <table align="center">
+        <tr>
+            <td style="padding-right: 20px"><a href="main"><b>Главная</b></a></td>
+            <td style="padding-right: 20px"><a href="cart"><b>Корзина</b></a></td>
+            <td style="padding-right: 20px"><a href="catalog"><b>Каталог</b></a></td>
+            <td style="padding-right: 20px"><a href="order"><b>Оформление заказа</b></a></td>
+        </tr>
+    </table>
+</div>
